% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ledpred.R
\name{scoreData}
\alias{scoreData}
\title{Predicting new regulatory regions}
\usage{
scoreData(data = NULL, data.granges = NULL, model, score.file = NULL)
}
\arguments{
\item{data}{data.frame containing the test set. This test set must have the same descriptive features as the one that were used to build the model.}

\item{data.granges}{Bioconductor GenomicRanges object containing the training set}

\item{model}{the SVM model}

\item{score.file}{A character string that will be used as the file name for the output file, if it is NULL (default), no file is writen}
}
\value{
A 2-columns dataframe. First column containg the SVM model prediction probabilities and the second containing the corresponding regions
}
\description{
\code{scoreData} function predict new regulatory regions using SVM model from a test data set
}
\examples{
data(seq.features)
data(svm.model)
pred.test <- scoreData(data.granges=seq.features, model=svm.model,
 score.file="test_prediction.tab")
}

